update: Retargeted the package to .NET Standard 2.1

+++ src/IPLocate/Exceptions/IPLocateApiKeyException.cs
@@ -5,7 +5,7 @@ namespace IPLocate.Exceptions
 	public class IPLocateApiKeyException : IPLocateApiException
 	{
 		public IPLocateApiKeyException(string message) 
-			: base(message, (int)HttpStatusCode.Forbidden)
+			: base(message, HttpStatusCode.Forbidden)
 		{
 		}
 	}
