+++ src/IPLocate/Exceptions/IPLocateNotFoundException.cs
@@ -5,7 +5,7 @@ namespace IPLocate.Exceptions
 	public class IPLocateNotFoundException : IPLocateApiException
 	{
 		public IPLocateNotFoundException(string message) 
-			: base(message, (int)HttpStatusCode.NotFound)
+			: base(message, HttpStatusCode.NotFound)
 		{
 		}
 	}
